third_party/blink/web_tests/platform/generic/fast/css/invalidation/clears-invalidation-whole-tree-expected.txt

Tests that a style recalc on the inner element after the outer one works correctly. See also crbug.com/346928.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(outer).width is "200px"
PASS getComputedStyle(inner).width is "150px"
PASS successfullyParsed is true

TEST COMPLETE
